Разбить код на модули
